# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/v1/enterprise": {
            "get": {
                "description": "Lista todas as empresas com suporte a paginação",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 - Empresa"
                ],
                "summary": "Listar todas as empresas",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número da página",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/enterprise.EnterpriseResponseDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria uma nova empresa com CNPJ e nome",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 - Empresa"
                ],
                "summary": "Criar empresa",
                "parameters": [
                    {
                        "description": "Dados da empresa",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enterprise.CreateEnterpriseDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/enterprise.EnterpriseResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
        "/admin/v1/enterprise/cnpj/{cnpj}": {
            "get": {
                "description": "Retorna os dados de uma empresa com base no CNPJ fornecido.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 - Empresa"
                ],
                "summary": "Buscar empresa por CNPJ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CNPJ da empresa",
                        "name": "cnpj",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/enterprise.EnterpriseResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza os dados de uma empresa a partir do CNPJ fornecido na URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 - Empresa"
                ],
                "summary": "Atualizar empresa por CNPJ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CNPJ da empresa atual",
                        "name": "cnpj",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados de atualização",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enterprise.UpdateEnterpriseByCNPJDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/enterprise.EnterpriseUpdatedResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove uma empresa com base no CNPJ fornecido.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 - Empresa"
                ],
                "summary": "Deletar empresa por CNPJ",
                "parameters": [
                    {
                        "type": "string",
                        "name": "cnpj",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/enterprise.EnterpriseDeletedResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
        "/admin/v1/enterprise/id/{id}": {
            "get": {
                "description": "Retorna os dados de uma empresa com base no ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 - Empresa"
                ],
                "summary": "Buscar empresa por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da empresa",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/enterprise.EnterpriseResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
        "/admin/v1/enterprise/nome/{nome}": {
            "get": {
                "description": "Retorna os dados de uma empresa com base no nome fornecido.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 - Empresa"
                ],
                "summary": "Buscar empresa por nome",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da empresa",
                        "name": "nome",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/enterprise.EnterpriseResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
        "/admin/v1/rules": {
            "get": {
                "description": "Retorna todas as regras com suporte a paginação",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 - Regra"
                ],
                "summary": "Listar regras",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Quantidade de resultados",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Deslocamento dos resultados",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rule.AdminRuleListDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
        "/admin/v1/rules/{id}": {
            "get": {
                "description": "Retorna os dados de uma regra com base no ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 - Regra"
                ],
                "summary": "Buscar regra por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da regra",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rule.AdminRuleDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
        "/admin/v1/rules/{id}/permissions": {
            "get": {
                "description": "Retorna a lista de permissões em formato namespace (ex: admin.enterprise.create)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 - Regra"
                ],
                "summary": "Buscar permissões da regra",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da regra",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/rule.RulePermissionDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        },
        "/admin/v1/user": {
            "post": {
                "description": "Cria um novo usuário no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v1 - Usuário"
                ],
                "summary": "Criar usuário",
                "parameters": [
                    {
                        "description": "Dados do usuário",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.AdminUserCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user.UserResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest_err.RestErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "enterprise.CreateEnterpriseDTO": {
            "type": "object",
            "required": [
                "cnpj",
                "nome"
            ],
            "properties": {
                "cnpj": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "enterprise.EnterpriseDeletedResponseDTO": {
            "type": "object",
            "properties": {
                "cnpj": {
                    "type": "string"
                }
            }
        },
        "enterprise.EnterpriseResponseDTO": {
            "type": "object",
            "properties": {
                "cnpj": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "enterprise.EnterpriseUpdatedResponseDTO": {
            "type": "object",
            "properties": {
                "cnpj": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "enterprise.UpdateEnterpriseByCNPJDTO": {
            "type": "object",
            "required": [
                "cnpj"
            ],
            "properties": {
                "cnpj": {
                    "description": "novo CNPJ opcional",
                    "type": "string"
                },
                "nome": {
                    "description": "nome opcional",
                    "type": "string"
                }
            }
        },
        "rest_err.Causes": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "rest_err.RestErr": {
            "type": "object",
            "properties": {
                "causes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest_err.Causes"
                    }
                },
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "rule.AdminRuleDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "rule.AdminRuleListDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rule.AdminRuleDTO"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "rule.RulePermissionDTO": {
            "type": "object",
            "properties": {
                "namespace": {
                    "type": "string"
                }
            }
        },
        "user.AdminUserCreateDTO": {
            "type": "object",
            "required": [
                "email",
                "enterprise_id",
                "nome",
                "numero",
                "rule_id",
                "senha"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "enterprise_id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string",
                    "minLength": 3
                },
                "numero": {
                    "type": "string"
                },
                "rule_id": {
                    "type": "integer"
                },
                "senha": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "user.UserResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "enterprise_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "numero": {
                    "type": "string"
                },
                "rule_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}